+++ preload.js
@@ -95,6 +95,8 @@ window.electronAPI = {
   clipboardWrite: text => ipcRenderer.invoke('app:clipboard', text),
   openExternal: url => ipcRenderer.invoke('app:openExternal', url),
   updateCheck: () => ipcRenderer.invoke('update:check'),
+  updateDownloadMsi: opts => ipcRenderer.invoke('update:downloadMsi', opts),
+  updateApplyMsi: opts => ipcRenderer.invoke('update:applyMsi', opts),
   appVersion: () => ipcRenderer.invoke('app:version'),
   appContributors: () => ipcRenderer.invoke('app:contributors'),